Create components for all homepage interface parts
